requirements.txt
appnope==0.1.0
backcall==0.2.0
backoff==1.8.0
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
ciso8601==2.1.3
decorator==4.4.2
deprecation==2.1.0
idna==2.10
ipdb==0.13.3
ipdf==0.0.1
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jsonpickle==0.9.6
jsonschema==2.6.0
methodtools==0.4.2
msgpack==1.0.0
packaging==20.4
parso==0.7.1
pendulum==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2018.4
pytzdata==2020.1
requests==2.24.0
simplejson==3.11.1
singer-python==5.9.0
six==1.15.0
-e git+https://github.com/nehiljain/target-csv.git@b68ae1364f43b80abd922f276f63c991533b976c#egg=target_csv
traitlets==5.0.4
tzlocal==2.1
urllib3==1.25.10
wcwidth==0.2.5
wirerope==0.4.2
